Create ElectronicsShop.java
A person wants to determine the most expensive computer keyboard and USB drive that can be purchased with a give budget. Given price lists for keyboards and USB drives and a budget, find the cost to buy them. If it is not possible to buy both items, return -1.

ex:
b=60;
keyboards = [40,50,60]
drivers= [5,8,12]

The person can buy a 40 keyboards + 12 USB drive =52  , or a 50 keyboards + 8 USB drive =58. Choose the latter as the more expensive option and return 58.
Function Description

Complete the getMoneySpent function in the editor below.

getMoneySpent has the following parameter(s):

int keyboards[n]: the keyboard prices
int drives[m]: the drive prices
int b: the budget
Returns

int: the maximum that can be spent, or -1 if it is not possible to buy both items